requirements.txt
Flask==3.0.2
python-dotenv
geocoder
requests
apscheduler
Flask-SQLAlchemy